move LoginUseCaseFactory to production code
