+++ test/core/navigator_factory/login_composer/login_use_case_factory_test.dart
@@ -1,12 +1,9 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:get/get.dart';
-import 'package:http/http.dart' as http;
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:thepos/core/auth_manager.dart';
-import 'package:thepos/core/config.dart';
+import 'package:thepos/core/login_composer/login_use_case_factory.dart';
 import 'package:thepos/core/login_composer/login_use_case_output_composer.dart';
 import 'package:thepos/core/navigator/navigator_factory.dart';
-import 'package:thepos/features/login/data/login_service/api_login/api_login_service.dart';
 import 'package:thepos/features/login/data/login_service/login_service.dart';
 import 'package:thepos/features/login/data/login_use_case/login_use_case.dart';
 import 'package:thepos/features/login/data/login_use_case/login_use_case_output.dart';
@@ -14,29 +11,7 @@ import 'package:thepos/features/login/presentation/controller/login_controller.d
 import 'package:thepos/features/login/presentation/login_router.dart';
 
 import '../../../helpers/navigator/navigator_factory_spy.dart';
-
-class LoginUseCaseFactory {
-  LoginUseCase makeUseCase(
-      SharedPreferences sharedPreferences, NavigatorFactory navigatorFactory) {
-    final Uri loginUri =
-        Uri.https(domain, '/mocks/thepos/thepos:v2/8473374/api/v2/login');
-    final LoginService loginService = ApiLoginService(http.Client(), loginUri);
-
-    final LoginController loginController = Get.put(LoginController());
-    final AuthManager authManager = AuthManager(sharedPreferences);
-    final LoginRouter router = LoginRouter(navigatorFactory);
-
-    final LoginUseCase useCase = LoginUseCase(
-      loginService: loginService,
-      output: LoginUseCaseOutputComposer(<LoginUseCaseOutput>[
-        loginController,
-        authManager,
-        router,
-      ]),
-    );
-    return useCase;
-  }
-}
+import 'helpers/dummy_login_service.dart';
 
 void main() {
   test('makeUseCase should compose correct outputs with correct order',
@@ -46,11 +21,17 @@ void main() {
     final SharedPreferences sharedPreferences =
         await SharedPreferences.getInstance();
 
+    final LoginService loginService = DummyLoginService();
+
+    final NavigatorFactory navigatorFactory = NavigatorFactorySpy();
+
     final LoginUseCaseFactory sut = LoginUseCaseFactory();
     final LoginUseCase useCase = sut.makeUseCase(
-      sharedPreferences,
-      NavigatorFactorySpy(),
+      loginService: loginService,
+      sharedPreferences: sharedPreferences,
+      navigatorFactory: navigatorFactory,
     );
+
     final LoginUseCaseOutput output = useCase.output;
 
     expect(output, isInstanceOf<LoginUseCaseOutputComposer>());
